Set up a scoreboard for the game
The scoreboard runs using an array. The outcome of each round is stored to a variable, is passed to the array and then a prompt confirms if user wants to check the score or not. If they respond in the affirmative, the array is iterated element wise, value of each element is checked and a counter containing the number of appearances of each type of element is finally used to display the score in the console.